Mouse buttons are now referred as names instead of numbers

+++ Scenes/UI/Settings/KeyBindEdit.cs
@@ -65,7 +65,7 @@ public partial class KeyBindEdit : Button
 		if (@event is InputEventMouse)
 		{
 			InputEventMouseButton button = @event as InputEventMouseButton;
-			bind = "Mouse" + (int)button.ButtonIndex;
+			bind = "Mouse" + button.ButtonIndex;
 		}
 
 		icon.key = bind;
